feat: Add --ignore-suffix flag to ignore any suffix
This commit introduces a new --ignore-suffix flag to the navidrome command. When this flag is used, the application will ignore the provided suffix when searching for tracks on Navidrome and DAB.

This feature is useful when syncing playlists from Spotify to Navidrome, as Spotify often has tracks with suffixes that are not present in Navidrome.

Fixes #8