feat: Add option to save album art
This commit introduces a new feature that allows users to save the album art as a `cover.jpg` file in the album's directory.

A new boolean option `saveAlbumArt` has been added to the configuration. When this option is set to `true`, the downloader will save the cover art for the downloaded album.

This resolves issue #9.